weather data fethed using latitude and longitude from google maps
